Important QOL changes! -Bugfixes -Added a new admin tool, get overmap object. This teleports overmap ships / planets / stations etc. to your location -Added a ship repairing tool (pending balance!) this lets you get your ships back to that factory new setting, just feed it batteries -Phasers now self-charge due to popular demand

+++ StarTrek13/Ships/overmap_combat.dm
@@ -217,6 +217,8 @@
 			target_turf = object
 		if((object in pilot.contents) || (object == mob))
 			return
+		if(!target_ship)
+			return
 		firinginprogress = TRUE
 		damage = SC.weapons.update_weapons()
 	else
@@ -231,7 +233,8 @@
 	damage = SC.weapons.update_weapons()
 	firinginprogress = FALSE
 	firecount = 0
-	qdel(current_beam) //Aka finish the attack, and ready the SFX for another one. This saves my eardrums :b1:
+	if(current_beam)
+		qdel(current_beam) //Aka finish the attack, and ready the SFX for another one. This saves my eardrums :b1:
 	current_beam = null
 	target_turf = null
